LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/static-router-fetch-event.https-expected.txt


FAIL Main resource matched the rule with fetch-event source assert_equals: expected 0 but got 1
PASS Subresource load matched the rule fetch-event source
